% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\docType{_PACKAGE}
\name{GLIMES}
\alias{GLIMES}
\title{Linear Mixed-Effects Model for single cell Expression Studies}
\description{
Single cell differential expression analysis for multiple samples
}
\author{
Chih-Hsuan Wu \email{chihhsuan@uchicago.edu}
}
