% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_methods.R
\name{poisson_glmm_DE}
\alias{poisson_glmm_DE}
\title{An implementation Poisson glmm DE method.}
\usage{
poisson_glmm_DE(
  sce,
  comparison,
  replicates,
  exp_batch = NULL,
  other_fixed = NULL,
  freq_expressed = 0.05
)
}
\arguments{
\item{sce}{a SingleCellExperiment object with raw count matrix.}

\item{comparison}{group comparison variable (e.g., conditions, celltype). The variable consists of exact two different groups.}

\item{replicates}{donor variable.}

\item{exp_batch}{experimental batches.}

\item{other_fixed}{donor variable.}

\item{freq_expressed}{a threshold for gene detection rate.}
}
\value{
A dataframe of Poisson glmm DE results with each row for a gene.
}
\description{
An implementation Poisson glmm DE method.
}
\examples{
data(Bcells_sce)
poisson_glmm_DE(Bcells_sce[1:10,], comparison = "stim", replicates = "ind")
}
